chore(rust): clean `lib.rs` for `ockam_transport_tcp`
