+++ implementations/rust/ockam/ockam_transport_tcp/src/transport.rs
@@ -7,7 +7,9 @@ use ockam_transport_core::TransportError;
 
 use crate::portal::TcpInletListenProcessor;
 use crate::workers::{ConnectionRole, TcpListenProcessor, TcpSendWorker};
-use crate::{parse_socket_addr, TcpOutletListenWorker, TcpRegistry};
+use crate::{TcpOutletListenWorker, TcpRegistry};
+
+pub(crate) const CLUSTER_NAME: &str = "_internals.transport.tcp";
 
 /// High level management interface for TCP transports
 ///
@@ -328,5 +330,66 @@ impl TcpTransport {
     pub async fn stop_outlet(&self, addr: impl Into<Address>) -> Result<()> {
         self.ctx.stop_worker(addr).await?;
         Ok(())
+    }
+}
+
+fn parse_socket_addr(s: &str) -> Result<SocketAddr> {
+    Ok(s.parse().map_err(|_| TransportError::InvalidAddress)?)
+}
+
+#[cfg(test)]
+mod test {
+    use core::fmt::Debug;
+    use ockam_core::{Error, Result};
+    use ockam_transport_core::TransportError;
+
+    use crate::transport::parse_socket_addr;
+
+    fn assert_transport_error<T>(result: Result<T>, error: TransportError)
+    where
+        T: Debug,
+    {
+        let invalid_address_error: Error = error.into();
+        assert_eq!(result.unwrap_err().code(), invalid_address_error.code())
+    }
+
+    #[test]
+    fn test_parse_socket_address() {
+        let result = parse_socket_addr("hostname:port");
+        assert!(result.is_err());
+        assert_transport_error(result, TransportError::InvalidAddress);
+
+        let result = parse_socket_addr("example.com");
+        assert!(result.is_err());
+        assert_transport_error(result, TransportError::InvalidAddress);
+
+        let result = parse_socket_addr("example.com:80");
+        assert!(result.is_err());
+        assert_transport_error(result, TransportError::InvalidAddress);
+
+        let result = parse_socket_addr("127.0.0.1");
+        assert!(result.is_err());
+        assert_transport_error(result, TransportError::InvalidAddress);
+
+        let result = parse_socket_addr("127.0.0.1:port");
+        assert!(result.is_err());
+        assert_transport_error(result, TransportError::InvalidAddress);
+
+        let result = parse_socket_addr("127.0.1:80");
+        assert!(result.is_err());
+        assert_transport_error(result, TransportError::InvalidAddress);
+
+        let result = parse_socket_addr("127.0.0.1:65536");
+        assert!(result.is_err());
+        assert_transport_error(result, TransportError::InvalidAddress);
+
+        let result = parse_socket_addr("127.0.0.1:0");
+        assert!(result.is_ok());
+
+        let result = parse_socket_addr("127.0.0.1:80");
+        assert!(result.is_ok());
+
+        let result = parse_socket_addr("127.0.0.1:8080");
+        assert!(result.is_ok());
     }
 }
